+++ Core/KernelSML/src/sml_ListenerThread.cpp
@@ -38,7 +38,7 @@ void ListenerThread::Run()
         sml::PrintDebug("Failed to create the internet listener socket.  Shutting down listener thread.") ;
         return ;
     }
-    
+
 #ifdef ENABLE_LOCAL_SOCKETS
     ok = m_LocalListenerSocket.CreateListener(m_Port, true);
     if (ok)
@@ -52,7 +52,7 @@ void ListenerThread::Run()
         return ;
     }
 #endif
-    
+
 #ifdef ENABLE_NAMED_PIPES
     ok = m_ListenerNamedPipe.CreateListener(m_Port) ;
     if (ok)
@@ -66,22 +66,22 @@ void ListenerThread::Run()
         return ;
     }
 #endif
-    
+
     m_Port = m_ListenerSocket.GetPort();
-    
+
     while (!m_QuitNow)
     {
         //sml::PrintDebug("Check for incoming connection") ;
-        
+
         // Check for an incoming client connection
         // This doesn't block.
         Socket* pSocket = m_ListenerSocket.CheckForClientConnection() ;
-        
+
         Socket* pLocalSocket = 0;
 #ifdef ENABLE_LOCAL_SOCKETS
         pLocalSocket = m_LocalListenerSocket.CheckForClientConnection();
 #endif
-        
+
 #ifdef ENABLE_NAMED_PIPES
         NamedPipe* pNamedPipe = m_ListenerNamedPipe.CheckForClientConnection();
 #endif
@@ -89,7 +89,7 @@ void ListenerThread::Run()
         {
             CreateConnection(pSocket);
         }
-        
+
         if (pLocalSocket)
         {
             CreateConnection(pLocalSocket);
@@ -99,7 +99,7 @@ void ListenerThread::Run()
         {
             CreateConnection(pNamedPipe);
             ok = m_ListenerNamedPipe.CreateListener(m_Port) ;
-            
+
             if (!ok)
             {
                 sml::PrintDebug("Failed to create the listener pipe.  Shutting down thread.") ;
@@ -107,29 +107,29 @@ void ListenerThread::Run()
             }
         }
 #endif
-        
+
         // Sleep for a little before checking for a new connection
         // New connections will come in very infrequently so this doesn't
         // have to be very rapid.
         sml::Sleep(0, 50) ;
     }
-    
+
     // Shut down our listener socket
     m_ListenerSocket.Close() ;
 #ifdef ENABLE_LOCAL_SOCKETS
     m_LocalListenerSocket.Close();
-    
+
 // This code shouldn't be necessary since the we unlink existing files on start
 // and since the file is in the user's home, permissions for that won't be an issue
     /*
         char f_name[256];
-        sprintf( f_name, "%s%d", sock::GetLocalSocketDir().c_str(), m_Port );
+        sprintf( f_name, "%s%d", sock::GetLocalSocketDir().c_str(), static_cast<int64_t>(m_Port));
         sml::PrintDebugFormat( "Attempting to delete %s", f_name );
         int del_status = unlink( f_name );
         if ( del_status == 0 )
             sml::PrintDebug( "Delete succeeded!" );
         else
-            sml::PrintDebugFormat( "Error occurred during delete attempt, error code %d", errno );
+            sml::PrintDebugFormat( "Error occurred during delete attempt, error code %d", static_cast<int64_t>(errno));
     */
 #endif
 #ifdef ENABLE_NAMED_PIPES
@@ -140,24 +140,24 @@ void ListenerThread::Run()
 void ListenerThread::CreateConnection(DataSender* pSender)
 {
     //sml::PrintDebugFormat("Got new connection on %s", pSender->GetName().c_str()) ;
-    
+
     // Create a new connection object for this socket
     Connection* pConnection = Connection::CreateRemoteConnection(pSender) ;
-    
+
     // Record our kernel object with this connection.  I think we only want one kernel
     // object even if there are many connections (because there's only one kernel) so for now
     // that's how things are set up.
     pConnection->SetUserData(m_pKernel) ;
-    
+
     // For debugging record that this is on the kernel side
     pConnection->SetIsKernelSide(true) ;
-    
+
     // Register for "calls" from the client.
     pConnection->RegisterCallback(ReceivedCall, NULL, sml_Names::kDocType_Call, true) ;
-    
+
     // Set up our tracing state to match the user's preference
     pConnection->SetTraceCommunications(m_pKernel->IsTracingCommunications()) ;
-    
+
     // Record the new connection in our list of connections
     m_pKernel->AddConnection(pConnection) ;
 }
